Juntar a AccountController com a UserController. - closed #27
