--- ItaLog/ItaLog.Domain/Interfaces/Repositories/IRoleRepository.cs
@@ -1,8 +1,0 @@
-﻿using ItaLog.Domain.Models;
-
-namespace ItaLog.Domain.Interfaces.Repositories
-{
-    public interface IRoleRepository : IRepositoryBase<Role>
-    {
-    }
-}
